+++ ZEA.Communications.Messaging.MassTransit/Interfaces/IDeadLetterConfigurator.cs
@@ -1,0 +1,6 @@
+namespace ZEA.Communications.Messaging.MassTransit.Interfaces;
+
+/// <summary>
+/// Base interface for dead-letter configurators.
+/// </summary>
+public interface IDeadLetterConfigurator;