+++ CHANGELOG.md
@@ -21,6 +21,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - **ZEA.Architecture.Pattern.Mediator.Abstractions.Responses**
     - Marked all the response classes as `[Obsolete]` to encourage migration to the new `LogicalError` classes.
+- **ZEA.Communications.Messaging.MassTransit**
+    - Marked the `RabbitMqBuilder` and `AzureServiceBusBuilder` classes as `[Obsolete]` to encourage migration to the
+      new `MassTransitBuilder` class.
 
 ### Renamed